Updating the style.css file and the app.js file.
Styling the navigation bar through the style.css file.
Also needed to update the app.js file because the background colour for the navigation bar is in that file.
Added href link for the brand in the navigation bar so that clicking on the 'SMC Tuckshop' text on the left will send the user to the home page.