Final Editing.
Edited the content of text in the home page.
Added the animated footer wave.
Editing the slides of the pages that use the bootstrap carousel.
Added the menu page showing the tuckshop menu.
Added more space between the update account form box and the footer in the Account page.
Changed the emails listed in the Contact page to the proper emails instead of test@my.sanctamaria.school.nz.